// lib/models/Session.ts
import mongoose, { Schema, Document } from "mongoose";

export interface ISession extends Document {
  _id: mongoose.Types.ObjectId;
  sessionId: string; // generated by SDK
  visitorId: string;
  websiteId: mongoose.Types.ObjectId;
  startedAt: Date;
  endedAt: Date;
  referrer: string;
  referrerDomain: string;
  utmSource: string;
  utmMedium: string;
  utmCampaign: string;
  duration: number; // in seconds
  bounce: boolean;
}

const SessionSchema = new Schema<ISession>(
  {
    sessionId: {
      type: String,
      required: true,
      unique: true,
      index: true,
    },
    visitorId: {
      type: String,
      required: true,
      index: true,
    },
    websiteId: {
      type: Schema.Types.ObjectId,
      ref: "Website",
      required: true,
      index: true,
    },
    startedAt: {
      type: Date,
      default: Date.now,
    },
    endedAt: {
      type: Date,
      default: Date.now,
    },
    referrer: {
      type: String,
      default: "",
    },
    referrerDomain: {
      type: String,
      default: "Direct",
    },
    utmSource: {
      type: String,
      default: "",
    },
    utmMedium: {
      type: String,
      default: "",
    },
    utmCampaign: {
      type: String,
      default: "",
    },
    duration: {
      type: Number,
      default: 0,
    },
    bounce: {
      type: Boolean,
      default: true,
    },
  },
  { timestamps: true }
);

export const Session = mongoose.models.Session || mongoose.model<ISession>("Session", SessionSchema);
export default Session;